requirements.txt
# Core dependencies
-e .  # This line installs the package in editable mode
streamlit>=1.32.0
langchain>=0.1.12
langchain-community>=0.0.27
langchain-core>=0.1.31
langgraph>=0.0.26
pydantic>=2.6.4
python-dotenv>=1.0.1
typing-extensions>=4.10.0
pypdf>=4.0.2

# Vector store dependencies
faiss-cpu>=1.7.4; platform_system!="Darwin" or platform_machine!="arm64"
faiss-cpu>=1.7.4; platform_system=="Darwin" and platform_machine=="arm64"
torch>=2.2.0  # Required by sentence-transformers and FAISS
transformers==4.37.2  # Required by sentence-transformers
sentence-transformers>=2.5.1
numpy>=1.26.4

# Development dependencies
pytest>=8.1.1
pytest-mock>=3.12.0
black>=24.3.0
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
types-requests==2.31.0.20240106
reportlab

# System dependencies (for M1 Mac)
# Note: These need to be installed via Homebrew:
# brew install swig cmake
ollama>=0.1.6